Make instrumenter_new() work by passing right env
The instrumenter_new/1 NIF (which wraps sky_instrumenter_new()) was
working in the sense that it didn't fail, but it was logging that it
could not load the Skylight from the environment. Apparently, this was
caused by a missing SKYLIGHT_AUTHENTICATION token in the environment
passed to it. It should be fixed now, even if I'm not sure the daemon is
being started yet.