Prevent segfaults by raising if traces are NULL
Functions like `Instrumenter.submit_trace/2` cause the trace to be freed
right away (`sky_instrumenter_submit_trace()` does that); this causes
segfaults if we try to do stuff with a trace after it's submitted. To
avoid this, I added a `CHECK_TRACE` macro to the C code so that whenever
we get a trace, we raise if it's NULL (without segfaulting).