Define the Skylight repo inside the existing repo
The Skylight repo is now still a separate repo that wraps the original
repository, but it's not defined separately from the original
repository. Instead, `use Skylight.Ecto.Repo` has to be called inside
the module that calls `use Ecto.Repo`; the Skylight repo wrapper will be
automatically defined (by appending `.Skylight` to the name of the
calling module).